refactor: address review and strictly align tests with architecture
* Removed the sprawling `test_tensor_manifest.py` file to adhere to the monolithic test organization defined in AGENTS.md.
* Migrated deterministic sorting validations for `PeftAdapterContract`, `MCPServerBindingProfile`, `SteadyStateHypothesisState`, and `NeuralAuditAttestationReceipt` into `test_ontology_validators.py`.
* Converted the tensor boundaries test into an exhaustive, property-based Hypothesis fuzzing test instead of using hardcoded synthetic edge cases.
* Performed true dead code elimination on `NDimensionalTensorManifest` by deleting redundant `isinstance` checks, as strict Pydantic parsing guarantees coercion.
* Correctly implemented the Topological Exemption on `ExecutionSpanReceipt` by removing the temporal sorting logic and declaring the sequence immutable via comment anchoring.
